requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
duckduckgo-search==6.2.6
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
jinja2==3.1.4
